I switched from the json parsers provided by Seaside to the JSON package from squeaksource.  I also specified the timestamps as seconds since the Unix epoch because Riak's map reduce parsing infrastructure can't handle iso8601 formatted timestamps at this time.
There is an add-on that can do it here:  https://github.com/basho/riak_function_contrib/blob/master/mapreduce/js/iso8601.js
but since its not in the default Riak installation I'm not putting it in as a dependency. 

The #testCountMessageTypes provides an example of how to run a map/reduce job on Riak.  It works great the first time and starts failing if you attempt to run it too frequently.  I have no idea why.